add "as" keyword and write docs
