+++ materialSync/material_cleanup.ts
@@ -22,9 +22,6 @@ yarn run cleanup
 
 const klassen = Object.keys(configs);
 
-fs.rmSync('src/pages/secure', { recursive: true, force: true });
-fs.rmSync('static/secure', { recursive: true, force: true });
-
 klassen.forEach((klass) => {
     const config = configs[klass];
     const tmp_dir = `versioned_docs/version-${klass}/.tmp`;
